+++ data-raw/states/ISD/prepare-ISD.R
@@ -2,53 +2,57 @@
 
 # This is a template for importing, cleaning, and exporting data
 # ready for the qPackage.
-library(qData)
-library(tidyverse)
 
 # Stage one: Collecting data
-ISD <- read.csv("data-raw/states/ISD/ISD_Version1_Dissemination.csv")
+ISD <- utils::read.csv("data-raw/states/ISD/ISD_Version1_Dissemination.csv")
 
 # Stage two: Correcting data
 # In this stage you will want to correct the variable names and
 # formats of the 'ISD' object until the object created
 # below (in stage three) passes all the tests. 
-
-ISD <- as_tibble(ISD) %>%
+ISD <- tibble::as_tibble(ISD) %>%
   dplyr::rename(Finish = End) %>% 
   # Renaming the end date column to avoid self reference in transmutate.
-  transmutate(ID = `COW.ID`,
-              Beg = standardise_dates(Start),
-              End = standardise_dates(Finish),
-              Label = standardise_titles(as.character(State.Name)),
-              COW_Nr = standardise_titles(as.character(COW.Nr.))) %>%
+  qData::transmutate(ID = `COW.ID`,
+                     Beg = qCreate::standardise_dates(lubridate::dmy(Start)),
+                     End = qCreate::standardise_dates(lubridate::dmy(Finish)),
+                     Label = qCreate::standardise_titles(as.character(State.Name)),
+                     COW_Nr = qCreate::standardise_titles(as.character(COW.Nr.))) %>%
   # Standardising the dummie variables
-  dplyr::mutate_at(vars(Micro, New.State),  ~replace(. == "", 0, .)) %>% 
-  dplyr::mutate_at(vars(Micro, New.State), ~replace(. == "x", 1, .)) %>%  
+  dplyr::mutate(across(c(Micro, New.State),  ~ replace(., . == "", 0))) %>% 
+  dplyr::mutate(across(c(Micro, New.State), ~ replace(., . == "X", 1))) %>%  
   #Dropping certain unnecessary columns.
   dplyr::select(-X, -X.1, -X.2, -X.3, -X.4, -X.5,  -X.6, -X.7) %>% 
   # Arranging dataset
-  dplyr::relocate(COW_Nr, ID, Beg, End, Label, Micro, New.State) %>%
+  dplyr::relocate(ID, Beg, End, COW_Nr, Label, Micro, New.State) %>%
   dplyr::arrange(Beg, ID)
-
-# qData includes several functions that should help cleaning and standardizing your data.
+# We know that COW data for "old" states is set to 1816-01-01 by default.
+# This is a rather uncretain date, that is, the dataset considers them states
+# on 1st January 1816, but they may have been established (much) earlier.
+# Let's signal to this uncretainty using `standardise_dates()` is a wrapper
+# for the `{messydates}` package which is designed to deal with date uncretianty.
+ISD$Beg <- qCreate::standardise_dates(stringr::str_replace_all(ISD$Beg,
+                                                               "1816-1-1|1816-01-1|1816-1-01|1816-01-01",
+                                                               "..1816-01-01"))
+# We can do the same for End dates to signal uncertainty. 
+ISD$End <- qCreate::standardise_dates(stringr::str_replace_all(ISD$End,
+                                                               "2011-12-31",
+                                                               "2011-12-31.."))
+# qData and qCreate include several other
+# functions that should help cleaning and
+# standardizing your data.
 # Please see the vignettes or website for more details.
-
-# #Additional Identifiers and Unicode Country Flags by using the countrycode package (https://github.com/vincentarelbundock/countrycode)
-# library(countrycode)
-# ID <- data.frame(ISD$ID, UnicodeSymb = countrycode::countrycode(ISD$ID, 'cowc', 'unicode.symbol'), ISO3_ID = countrycode::countrycode(ISD$ID, 'cowc', 'iso3c'), EuroStat_code = countrycode::countrycode(ISD$ID, 'cowc', 'eurostat'), ECB_code = countrycode::countrycode(ISD$ID,'cowc', 'ecb'), IANA_TLD = countrycode::countrycode(ISD$ID, 'cowc', "cctld"), un_code = countrycode::countrycode(ISD$ID, 'cowc', 'un'),Continent = countrycode::countrycode(ISD$ID, 'cowc', 'continent'), EU = countrycode::countrycode(ISD$ID, 'cowc', 'eu28'), Currency = countrycode::countrycode(ISD$ID, 'cowc', 'iso4217c'))
-# ID  <- as_tibble(ID)
-# ISD <- dplyr::bind_cols(ISD, ID)
-# rm(ID)
-# # Issue, historic countries (Denmark for example) that have a non continuous existence are coded with the same ISO-ID's for both occurrences of existence.This also prevents inner joins.
-
 # Stage three: Connecting data
 # Next run the following line to make ISD available within the qPackage.
-export_data(ISD, database = "states", URL="http://www.ryan-griffiths.com/data")
+qCreate::export_data(ISD, database = "states",
+                     URL="http://www.ryan-griffiths.com/data")
+
 # This function also does two additional things.
-# First, it creates a set of tests for this object to ensure adherence to certain standards.
-# You can hit Cmd-Shift-T (Mac) or Ctrl-Shift-T (Windows) to run these tests locally at any point.
-# Any test failures should be pretty self-explanatory and may require you to return
-# to stage two and further clean, standardize, or wrangle your data into the expected format.
+# First, it creates a set of tests for this object to ensure adherence
+# to certain standards. You can hit Cmd-Shift-T (Mac) or Ctrl-Shift-T (Windows)
+# to run these tests locally at any point. Any test failures should be pretty
+# self-explanatory and may require you to return to stage two and further clean,
+# standardize, or wrangle your data into the expected format.
 # Second, it also creates a documentation file for you to fill in.
-# Please make sure that you cite any sources appropriately and fill in as much detail
-# about the variables etc as possible.
+# Please make sure that you cite any sources appropriately and fill in as
+# much detail about the variables etc as possible.