+++ tests/testthat/test_ISD.R
@@ -20,11 +20,17 @@ test_that("object has the correct variables", {
 })
 
 # Variables with dates are standardized
-test_that("dates are standardised", {
-  expect_col_is_date(states[["ISD"]], vars(Beg))
-  expect_col_is_date(states[["ISD"]], vars(End))
+test_that("Columns with dates are standardized", {
+  expect_equal(class(states[["ISD"]]$Beg), "messydt")
   expect_false(any(grepl("/", states[["ISD"]]$Beg)))
-  expect_false(any(grepl("/", states[["ISD"]]$End)))
+  expect_false(any(grepl("^[:alpha:]$",
+                         states[["ISD"]]$Beg)))
+  expect_false(any(grepl("^[:digit:]{2}$",
+                         states[["ISD"]]$Beg)))
+  expect_false(any(grepl("^[:digit:]{3}$",
+                         states[["ISD"]]$Beg)))
+  expect_false(any(grepl("^[:digit:]{1}$",
+                         states[["ISD"]]$Beg)))
 })
 
 # Labels are standardized